Refactor: Move channel endpoint validator to its own directory
