+++ apps/api/src/app/channel-endpoints/dtos/update-channel-endpoint-request.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { getApiPropertyExamples } from '@novu/application-generic';
 import { ChannelEndpointByType, ChannelEndpointType } from '@novu/shared';
 import { IsDefined } from 'class-validator';
-import { IsValidChannelEndpoint } from '../../shared/validators/channel-endpoint.validator';
+import { IsValidChannelEndpoint } from '../validators/channel-endpoint.validator';
 
 export class UpdateChannelEndpointRequestDto {
   @ApiProperty({
