+++ apps/api/src/app/channel-endpoints/validators/channel-endpoint.validator.ts
